requirements.txt

Django==3.0
djangorestframework==3.11.0
django-cors-headers==3.3.0
xlwt==1.3.0
